Do not wait until next update window when .spec changes
We are currently waiting until the next update window for even
considering to check if a reinstall/update could be done. So if a user
changes some parameter in the spec (ie, the version from `version: 1.*`
to `version: 2.*`) they will have to wait until maybe next Sunday 12pm
for seeing that change baing applied.

This change detects changes in the .spec in the AmbassadorInstallation
(using the annotation where Kuberneteas saves the last applied settings).
If a change is detected, changes are applied immediately.

Signed-off-by: Alvaro Saurin <alvaro.saurin@gmail.com>